Move `_mask_email` to fix an import cycle
